Move PagingTest to its own package, apply PagingTest to Coverage Report
Included via a command-line argument "runPagingTest"
